model_template/examples/dinov3_fewshot/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pyyaml>=6.0.0
pillow>=10.0.0
httpx>=0.27.0
numpy>=1.26.0
